+++ src/main/java/at/porscheinformatik/seleniumcomponents/ActiveSeleniumComponent.java
@@ -454,3 +454,4 @@ public interface ActiveSeleniumComponent extends SeleniumComponent {
         }
     }
 }
+