+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/ClarityNumberInputContainerComponent.java
@@ -1,13 +1,11 @@
 package at.porscheinformatik.seleniumcomponents.clarity;
 
-import static at.porscheinformatik.seleniumcomponents.WebElementSelector.selectByTagName;
-import static at.porscheinformatik.seleniumcomponents.WebElementSelector.selectByTestId;
-import static at.porscheinformatik.seleniumcomponents.WebElementSelector.selectByText;
+import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.SeleniumKeyboardShortcuts;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 import at.porscheinformatik.seleniumcomponents.component.InputComponent;
-import org.openqa.selenium.Keys;
 
 public class ClarityNumberInputContainerComponent extends ClarityFormControlContainer {
 
@@ -68,8 +66,8 @@ public class ClarityNumberInputContainerComponent extends ClarityFormControlCont
     }
 
     public void clear() {
-        // Mark the whole input text and delete it
-        input.sendKeys(Keys.chord(Keys.CONTROL, "A"), Keys.BACK_SPACE);
+        // Mark the whole input text and delete it, because clear() does not trigger the change detection in Angular.
+        input.sendKeys(SeleniumKeyboardShortcuts.deleteAll());
     }
 
     /**